pass through the snapshout source
