+++ crates/iota-config/Cargo.toml
@@ -14,6 +14,8 @@ clap.workspace = true
 csv.workspace = true
 dirs.workspace = true
 fastcrypto.workspace = true
+flate2.workspace = true
+iota-sdk = { version = "1.1", default-features = false, features = ["std"] }
 object_store.workspace = true
 once_cell.workspace = true
 prometheus.workspace = true
