+++ crates/iota-config/src/lib.rs
@@ -19,6 +19,7 @@ pub mod node;
 pub mod node_config_metrics;
 pub mod object_storage_config;
 pub mod p2p;
+pub mod snapshot;
 pub mod transaction_deny_config;
 
 use iota_types::multiaddr::Multiaddr;
