+++ crates/iota-genesis-builder/Cargo.toml
@@ -13,7 +13,6 @@ bigdecimal = "0.4"
 camino.workspace = true
 clap.workspace = true
 fastcrypto.workspace = true
-flate2.workspace = true
 fs_extra = "1.3"
 iota-sdk = { version = "1.1", default-features = false, features = ["irc_27", "irc_30", "std"] }
 itertools.workspace = true
